[REF] エラー処理の分離
